techpack: add tech package support
Enable tech packages to have kernel drivers that are physically
located in the kernel source tree but not present in the kernel
repository. This is done via a manifest line item to fetch a
techpackage-module into $KERNEL/techpack/$techpackage-module.
$KERNEL/techpack/Kbuild will automatically detect any
subdirectories and link them to the kernel's kbuild system.

The resulting layout of techpack within kernel source would be as
follows :-

    kernel/
    +-- techpack/
        +-- Kbuild
        +-- stub
        ¦    +-- Makefile
        ¦    +-- include
        ¦    ¦   +-- uapi
        ¦    ¦       +-- Kbuild
             +-- stub.c
        +-- techpackage-module
            +-- Makefile
            +-- include
            ¦   +-- uapi
            ¦       +-- Kbuild

$KERNEL/techpack only contains Kbuild (no Makefile) at the
toplevel and this Kbuild takes care of both compiling the
subdirectories and of exporting the needed header files therein.
The reason for having only Kbuild at the top is that Kbuild
and Makefile cannot exist together in same directory and Makefile
doesn't cater to uapi header installation.

stub is an empty techpackage-module which serves both as a sample
layout and satisfies the requirement of kernel build system, by
providing necessary buit-in.o, when no other techpackage-module has
been pulled under techpack/.

$KERNEL/techpack/techpackage-module should have a Makefile at the
top and Kbuild under $KERNEL/techpack/techpackage-module/include/uapi
directory. The uapi headers pertaining to a techpackage-module
should reside under $KERNEL/techpack/techpackage-module/include/uapi
and associated Kbuild should have necessary rules to export it.

Signed-off-by: Bryan Huntsman <bryanh@codeaurora.org>
Signed-off-by: Imran Khan <kimran@codeaurora.org>
Signed-off-by: engstk <eng.stk@sapo.pt>
Signed-off-by: YaoDao <jsjson@163.com>