Adding-Agile-SD-TCP-module-and-modifying-Kconfig-and-makefile
This commit implements a new TCP congestion control algorithm, namely Agile-SD.
A detailed description of Agile-SD is published in the following 2 articles:

[1] "Agile-SD: a Linux-based TCP congestion control algorithm for supporting high-speed and short-distance networks",
Alrshah, M.A., Othman, M., Ali, B. and Hanapi, Z.M. Journal of Network and Computer Applications, Vol. 55, pages.181-190, May-June 2015.

[2] "Modeling the Throughput of the Linux-Based Agile-SD Transmission Control Protocol",
Alrshah, M.A., Othman, M., Ali, B. and Hanapi, Z.M.
IEEE Access journal, Vol. 4, pages.9724-9732, Jan 2017.

The Internet has predominantly used Reno or CUBIC, relying on packet loss as the
signal to slow down. While this worked well for many years, these congestion control
algorithms are unfortunately unable to present an acceptable bandwidth utilization
over today's networks. On today's Internet, data losses are very common and the
existing congestion control algorithms are very sensitive to packet losses causing
unwanted delay and retransmission of data. Also, the existing congestion control
algorithms shows long epochs which affect the general performance of TCP.

Agile-SD has significantly increased throughput, reduced sensitivity to packet loss,
and improved the ability to work with small and large buffers. Agile-SD shrinks the
needed epoch time to recover after data loss, which improves throughput up to 50% in
many cases, see [1,2].

Agile-SD is a Congestion Control Algorithm for High-speed Networks. Agile-SD is a
loss-based and RTT-independent TCP congestion control algorithm designed to support
high-speed networks. Agile-SD requires only changes on the sender side, not in the
network or the receiver side. So, it can be incrementally deployed on today's Internet
and/or in datacenters. Agile-SD introduces the Agility Factor Mechanism (AFM), which
allows Agile-SD to deal with small buffer sizes while reducing its sensitivity to packet
loss. Due to the use of this mechanism, Agile-SD improves the throughput of TCP up to 50%
compared to Cubic-TCP and Compound-TCP in many cases, especially when the buffer is small
and the data loss is common. Its performance was evaluated using simulation and testbed
to measure the average throughput, loss ratio, and fairness.

Our long-term goal is to improve the congestion control algorithms used on the Internet.
We are hopeful that Agile-SD can help advance the efforts toward this goal, and motivate
the community to do further research.

Signed-off-by: Mohamed A. Alrshah <mohamed.a.alrshah@ieee.org>
Signed-off-by: Mohamed Othman <mothman@upm.edu.my>
Signed-off-by: Borhanuddin Ali <borhan@upm.edu.my>
Signed-off-by: Zurina Hanapi <zurinamh@upm.edu.my>
Signed-off-by: YaoDao <ETO>